+++ src/AnyStatus.Plugins/Widgets/PerformanceCounter/PerformanceCounter.cs
@@ -1,13 +1,12 @@
 ﻿using AnyStatus.API;
 using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
-using System.Diagnostics;
 
 namespace AnyStatus
 {
     [DisplayName("Performance Counter")]
     [Description("Experimental. Shows the value of a performance counter")]
-    public class PerformanceCounter : Metric, IMonitored
+    public class PerformanceCounter : Metric, ISchedulable
     {
         private const string Category = "Performance Counter";
 
@@ -37,32 +36,5 @@ namespace AnyStatus
         [Category(Category)]
         [Description("")]
         public string InstanceName { get; set; }
-    }
-
-    public class PerformanceCounterMonitor : IMonitor<PerformanceCounter>
-    {
-        [DebuggerStepThrough]
-        public void Handle(PerformanceCounter item)
-        {
-            if (string.IsNullOrWhiteSpace(item.MachineName))
-                item.MachineName = "localhost";
-
-            item.Value = (int)GetValue(item);
-
-            item.State = State.Ok;
-        }
-
-        public float GetValue(PerformanceCounter item)
-        {
-            var counter = new System.Diagnostics.PerformanceCounter
-            {
-                CategoryName = item.CategoryName,
-                CounterName = item.CounterName,
-                InstanceName = item.InstanceName,
-                MachineName = item.MachineName
-            };
-
-            return counter.NextValue();
-        }
     }
 }
