+++ src/AnyStatus.Plugins/packages.config
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="AnyStatus.API" version="2.0.17-alpha" targetFramework="net46" />
+  <package id="AnyStatus.API" version="2.0.18-alpha" targetFramework="net46" />
   <package id="Extended.Wpf.Toolkit" version="3.1" targetFramework="net452" />
   <package id="Microsoft.Web.Administration" version="7.0.0.0" targetFramework="net452" />
 </packages>
